backend/requirements.txt
Flask
deepface
Pillow
numpy
pymongo dnspython
flask-cors